Use C++11 like yast2-core does (boo#914255).
Strictly speaking we should pull in the CXX flags exported by
yast2-core, but the mechanism to pass that is already in a mess,
too big a yak to shave now.

Declaring that this project uses C++11 on its own right is also correct.

Using 'gnu++0x' (deprecated now) instead of 'gnu++11' because old GCC
we use on Travis does not support the latter.