change name to y2start.
Reasoning is that y2base is still needed for old testsuite that require
a lot of very specific y2base features like printing log to command line
or logging config file. So for this reason use ruby starter with
different name and drop y2base when old testsuite will be dropped.