fix calling command line interface (bsc#1033993)
This change require a bit longer explanantion. The bug is result of
two changes. The first one is switch to ruby script as start point of
YaST. The second one is bug mentioning that there is different title
after this change. I found that the second one is caused by work-around
in libyui which no longer works. So I decide to do it properly and
always set proper title when starting YaST. And here it hits a bug.
UI in YaST is lazy loaded, it means that it is created when first call
to UI namespace happen. And setting title is that first UI call, so it
creates UI even when command line is required. So this change change
title only if there are no arguments for client. So module is showed in
UI.