A missing `textdomain` declaration is optionally an exception (bsc#1130822)
This kind of error is a programmer error which should be caught very
early during the development or testing, so users should not see them.
BUT we may have allowed this kind of bugs to creep in over time
so it would be a nuisance to start failing in users faces now.
-> Make it opt-in for testing.

The exception is enabled by export Y2STRICTTEXTDOMAIN=1