Typed Builtins.find
Found corner cases (which were not type correct YCP) that produced weird
exceptions, made them explanatory TypeErrors.

Removed deep_copy on an Integer.

The sig lies a bit in that it allows FalseClass as the object argument.
If I omit it, the static type checker will complain that the "else" clause
after "when String, when Array" can never be reached.
That would be correct if the sig were in the RB file, but it is in the RBI
file, not used at run time.
I need to figure out the proper way to have sigs in RB and RBI and check that
they are in sync. Currently it does not work :(